PDFBOX-2147: Removed unnecessary boxing, unboxing; convert to foreach loop
git-svn-id: https://svn.apache.org/repos/asf/pdfbox/trunk@1613756 13f79535-47bb-0310-9956-ffa450edef68

+++ jempbox/src/main/java/org/apache/jempbox/xmp/XMPSchemaPhotoshop.java
@@ -338,7 +338,7 @@ public class XMPSchemaPhotoshop extends XMPSchema
     {
         if( s != null )
         {
-            if( s.intValue() < 1 || s.intValue() > 8 )
+            if( s < 1 || s > 8 )
             {
                 throw new RuntimeException( "Error: photoshop:Urgency must be between 1 and 8.  value=" + s );
             }
